Add option skip_on_error for creating buildcaches

+++ lib/spack/spack/binary_distribution.py
@@ -1186,6 +1186,7 @@ def _build_tarball(
     relative=False,
     unsigned=False,
     allow_root=False,
+    skip_on_error=False,
     key=None,
     regenerate_index=False,
 ):
@@ -1261,18 +1262,26 @@ def _build_tarball(
         try:
             make_package_relative(workdir, spec, allow_root)
         except Exception as e:
-            shutil.rmtree(workdir)
-            shutil.rmtree(tarfile_dir)
-            shutil.rmtree(tmpdir)
-            tty.die(e)
+            if skip_on_error:
+                tty.warn(e)
+                tty.warn("Skipping package since skip_on_error flag is set")
+            else:
+                shutil.rmtree(workdir)
+                shutil.rmtree(tarfile_dir)
+                shutil.rmtree(tmpdir)
+                tty.die(e)
     else:
         try:
             check_package_relocatable(workdir, spec, allow_root)
         except Exception as e:
-            shutil.rmtree(workdir)
-            shutil.rmtree(tarfile_dir)
-            shutil.rmtree(tmpdir)
-            tty.die(e)
+            if skip_on_error:
+                tty.warn(e)
+                tty.warn("Skipping package since skip_on_error flag is set")
+            else:
+                shutil.rmtree(workdir)
+                shutil.rmtree(tarfile_dir)
+                shutil.rmtree(tmpdir)
+                tty.die(e)
 
     # create gzip compressed tarball of the install prefix
     # On AMD Ryzen 3700X and an SSD disk, we have the following on compression speed:
@@ -1656,7 +1665,7 @@ def dedupe_hardlinks_if_necessary(root, buildinfo):
         buildinfo[key] = new_list
 
 
-def relocate_package(spec, allow_root):
+def relocate_package(spec):
     """
     Relocate the given package
     """
